processing/requirements.txt
pandas
matplotlib
wordcloud
pillow
requests
altair
altair_viewer
python-dotenv
